# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user": {
            "get": {
                "description": "get all information about me",
                "tags": [
                    "Users"
                ],
                "summary": "Get about me",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "UserModel"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "description": "delete user account",
                "tags": [
                    "Users"
                ],
                "summary": "delete account",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.UserModel"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "nil"
                        }
                    }
                }
            },
            "patch": {
                "description": "edit user information",
                "tags": [
                    "Users"
                ],
                "summary": "edit user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.UserModel"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "nil"
                        }
                    }
                }
            }
        },
        "/user/cred": {
            "delete": {
                "description": "delete webauthn credential",
                "tags": [
                    "Users"
                ],
                "summary": "delete credential",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.UserModel"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "nil"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Just Logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.UserModel"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "nil"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Channel": {
            "type": "object",
            "properties": {
                "banner": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "owerId": {
                    "type": "integer"
                },
                "owner": {
                    "$ref": "#/definitions/domain.UserModel"
                },
                "socialLink": {
                    "type": "string"
                },
                "subscribers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.UserModel"
                    }
                }
            }
        },
        "domain.Role": {
            "type": "object",
            "properties": {
                "channel": {
                    "$ref": "#/definitions/domain.Channel"
                },
                "channelId": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "permission": {
                    "type": "integer"
                },
                "user": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.UserModel"
                    }
                }
            }
        },
        "domain.UserModel": {
            "type": "object",
            "properties": {
                "credentials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/webauthn.Credential"
                    }
                },
                "disable": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "incredentials": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Role"
                    }
                },
                "subscribtion": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Channel"
                    }
                },
                "username": {
                    "type": "string"
                },
                "valideAccount": {
                    "type": "boolean"
                }
            }
        },
        "webauthn.Authenticator": {
            "type": "object",
            "properties": {
                "aaguid": {
                    "description": "The AAGUID of the authenticator. An AAGUID is defined as an array containing the globally unique\nidentifier of the authenticator model being sought.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "cloneWarning": {
                    "description": "CloneWarning - This is a signal that the authenticator may be cloned, i.e. at least two copies of the\ncredential private key may exist and are being used in parallel. Relying Parties should incorporate\nthis information into their risk scoring. Whether the Relying Party updates the stored signature\ncounter value in this case, or not, or fails the authentication ceremony or not, is Relying Party-specific.",
                    "type": "boolean"
                },
                "signCount": {
                    "description": "SignCount -Upon a new login operation, the Relying Party compares the stored signature counter value\nwith the new signCount value returned in the assertion’s authenticator data. If this new\nsignCount value is less than or equal to the stored value, a cloned authenticator may\nexist, or the authenticator may be malfunctioning.",
                    "type": "integer"
                }
            }
        },
        "webauthn.Credential": {
            "type": "object",
            "properties": {
                "attestationType": {
                    "description": "The attestation format used (if any) by the authenticator when creating the credential.",
                    "type": "string"
                },
                "authenticator": {
                    "description": "The Authenticator information for a given certificate",
                    "allOf": [
                        {
                            "$ref": "#/definitions/webauthn.Authenticator"
                        }
                    ]
                },
                "id": {
                    "description": "A probabilistically-unique byte sequence identifying a public key credential source and its authentication assertions.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "publicKey": {
                    "description": "The public key portion of a Relying Party-specific credential key pair, generated by an authenticator and returned to\na Relying Party at registration time (see also public key credential). The private key portion of the credential key\npair is known as the credential private key. Note that in the case of self attestation, the credential key pair is also\nused as the attestation key pair, see self attestation for details.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}